Add Flyway dependencies, update entity IDs to Integer, and create migration scripts for addresses, companies, city halls, and profiles

+++ src/main/java/uno/tek/birth_management_service/profiles/ProfilesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfilesRepository extends JpaRepository<Profile, Long> {
+public interface ProfilesRepository extends JpaRepository<Profile, Integer> {
     boolean existsByEmail(String email);
 
     boolean existsByPhone(String phone);
